Refactored SettingsManager and added callbacks
