+++ src/Scenes/Settings/AudioPanel.tscn
@@ -2,7 +2,7 @@
 
 [ext_resource type="Script" uid="uid://jdg343tpghf3" path="res://src/Scripts/Scenes/Settings/AudioPanel.cs" id="1_dl324"]
 
-[node name="AudioPanel" type="Control" unique_id=173584112]
+[node name="AudioPanel" type="Control" unique_id=173584112 node_paths=PackedStringArray("MasterVolumeSlider")]
 layout_mode = 3
 anchors_preset = 15
 anchor_right = 1.0
@@ -10,6 +10,7 @@ anchor_bottom = 1.0
 grow_horizontal = 2
 grow_vertical = 2
 script = ExtResource("1_dl324")
+MasterVolumeSlider = NodePath("AudioContainer/MVolumeContainer/MVolumeSlider")
 
 [node name="AudioContainer" type="VBoxContainer" parent="." unique_id=1107211280]
 layout_mode = 1
