+++ src/Scripts/Scenes/Settings/SettingsMap.cs
@@ -1,0 +1,27 @@
+public static class SettingsMap
+{
+	public static class Section
+	{
+		public const string AUDIO = "Audio";
+		public const string VIDEO = "Video";
+		public const string KEYBINDS = "Keybindings";
+	}
+
+	public static class Audio
+	{
+		public const string MASTER_VOLUME = "master_volume";
+	}
+
+	public static class Video
+	{
+		public const string FULLSCREEN = "fullscreen";
+	}
+
+	public static class Keys
+	{
+		public const string MOVE_LEFT = "move_left";
+		public const string MOVE_RIGHT = "move_right";
+		public const string MOVE_UP = "move_up";
+		public const string MOVE_DOWN = "move_down";
+	}
+}
