+++ src/Scripts/Scenes/Settings/VideoPanel.cs
@@ -1,14 +1,13 @@
-using System;
 using Godot;
 
 public partial class VideoPanel : Control
 {
-	public override void _Ready()
-	{
-	}
+	[Export]
+	private CheckBox FullscreenToggle { get; set; }
 
-	private void OnFullscreenToggled(bool toggledOn)
-	{
-		// SettingsManager.SetFullscreen();
-	}
+	public override void _Ready() =>
+		FullscreenToggle.ButtonPressed = (bool)SettingsManager.Instance.GetSetting(SettingsMap.Section.VIDEO, SettingsMap.Video.FULLSCREEN, false);
+
+	private static void OnFullscreenToggled(bool toggledOn) =>
+		SettingsManager.Instance.SetSetting(SettingsMap.Section.VIDEO, SettingsMap.Video.FULLSCREEN, toggledOn);
 }
