+++ src/Scripts/Scenes/SettingsMenu.cs
@@ -1,4 +1,3 @@
-using System;
 using Godot;
 
 public enum SettingsTab
